# app/static/swagger.yaml
openapi: 3.0.0
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /users:
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Returns all users in the cinema
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#components/schemas/UsersOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Create new user
      requestBody:
        required: true
        description: New user info
        content:
          application/json:
            schema:
              $ref: '#components/schemas/UserIn'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Input
          content:
            application/json:
              example:
                message: Please, specify "username", "name","email", "password" and "age".
  /users/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: id of the user
          required: true
      summary: Get one user by id
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '404':
          description: User Not Found
          content:
            application/json:
              example:
                message: User not found.
    patch:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Updates user's info
      parameters:
        - name: id
          in: path
          description: id of the user
          required: true
      requestBody:
        required: true
        description: User's updated info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Updated
        '404':
          description: User Not Found
          content:
            application/json:
              example:
                message: User not found.
    delete:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Remove user by id
      parameters:
        - name: id
          in: path
          description: id of the user
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: User was successfully deleted.
        '404':
          description: User Not Found
          content:
            application/json:
              example:
                message: User not found.
  /tickets:
    get:
      security:
        - bearerAuth: []
      tags:
        - Tickets
      summary: Returns all tickets that were purchased at this cinema
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#components/schemas/TicketsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - Tickets
      summary: Buy a ticket
      requestBody:
        required: true
        description: New ticket info
        content:
          application/json:
            schema:
              $ref: '#components/schemas/TicketIn'
      responses:
        '201':
          description: Successful Operation
          content:
            application/json:
              example:
                id: "{ticket.id}"
                seat: "{ticket.seat}"
        '200':
          description: The wrong place was chosen
          content:
            application/json:
              example:
                message: 'Please, choose another seat.'
        '400':
          description: Bad Input
          content:
            application/json:
              example:
                message: Please, specify "seat", "user_id" and "session_id".
  /mytickets:
    get:
      security:
        - bearerAuth: []
      tags:
        - Tickets
      summary: Returns all tickets that were purchased by the selected user
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#components/schemas/TicketsOut'
  /free_seats/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Tickets
      parameters:
        - name: id
          in: path
          description: id of the session
          required: true
      summary: Get available seats for session by id
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Available_seats'
        '400':
          description: Bad input
          content:
            application/json:
              example:
                message: Such session not exist. Please,try another one
        '404':
          description: Session Not Found
          content:
            application/json:
              example:
                message: Session not found.
  /tickets/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Tickets
      parameters:
        - name: id
          in: path
          description: id of the session
          required: true
      summary: Get  tickets by session id
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tickets_by_sess'
        '400':
          description: Bad input
          content:
            application/json:
              example:
                message: Tickets not found.
        '404':
          description: Session Not Found
          content:
            application/json:
              example:
                message: Tickets not found.
  /sold_tickets/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Tickets
      parameters:
        - name: id
          in: path
          description: id of the film
          required: true
      summary: Get  number of tickets that were sold for a particular movie
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              application/json:
                example:
                  message: Sold tickets for this film {result}
        '404':
          description: Bad input
          content:
            application/json:
              example:
                message: Tickets not found.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserIn:
      type: object
      properties:
        name:
          type: string
          example: testuser
        age:
          type: integer
          example: 30
        username:
          type: string
          example: testuser
        password:
          type: string
          example: testuser
        email:
          type: string
          example: testuser
        is_admin:
          type: bool
          example: false
    UserOut:
      type: object
      properties:
        id:
          type: integer
          example: 3
        name:
          type: string
          example: testuser
          age:
            type: integer
            example: 30
          username:
            type: string
            example: testuser
          password:
            type: string
            example: testuser
          email:
            type: string
            example: testuser
          is_admin:
            type: bool
            example: false
...
    Available_seats:
      type: array
      items:
        type: integer
        example: 1,2,3,4,5,6,7,8,9,10
    Tickets_by_sess:
      type: array
      items:
        $ref: '#components/schemas/TicketOut'